Add duke.core package
