+++ src/main/java/duke/core/Storage.java
@@ -1,3 +1,5 @@
+package duke.core;
+
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.IOException;
