+++ src/main/java/duke/core/Ui.java
@@ -1,3 +1,5 @@
+package duke.core;
+
 import java.util.List;
 import java.util.stream.Stream;
